+++ src/Stackmaster/PatchInstaller.cs
@@ -36,7 +36,7 @@ namespace Stackmaster
                 Postfix(typeof(Player), "HaveRequirementItems", new[] { typeof(Recipe), typeof(bool), typeof(int), typeof(int) }, typeof(NearbyRequirementPatches), "RecipePostfix"),
                 Postfix(typeof(Player), "HaveRequirements", new[] { typeof(Piece), typeof(Player.RequirementMode) }, typeof(NearbyRequirementPatches), "PiecePostfix"),
                 Postfix(typeof(Hud), "SetupPieceInfo", new[] { typeof(Piece) }, typeof(NearbyBuildHudPatch)),
-                Postfix(typeof(InventoryGui), "SetupRequirement", new[]
+                Both(typeof(InventoryGui), "SetupRequirement", new[]
                 {
                     typeof(UnityEngine.Transform), typeof(Piece.Requirement), typeof(Player), typeof(bool), typeof(int), typeof(int)
                 }, typeof(NearbyCraftingHudPatch)),